#10.1: refactor(render): extract NutrientManager into components/NutrientManager.js
