+++ package.json
@@ -4,9 +4,9 @@
   "description": "",
   "main": "main.js",
   "scripts": {
-  "test": "echo \"Error: no test specified\" && exit 1",
-  "test:unit": "vitest",
-  "dev": "vite",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "test:unit": "vitest",
+    "dev": "vite",
     "format": "prettier --write .",
     "lint": "eslint . --ext .js",
     "lint:fix": "npx eslint . --ext .js --config eslint.config.cjs --fix && npx prettier --write .",
